Code Louisville SQL project
I want to thank the instructors for taking the time to teach this class.  I got alot out of it.  Towards the end I was more focused on the project, that's why I wasn't as participating as much in the discussions.  Please let me know if you have any questions?  I have enclosed to files.  One file has the tables and the other table is my queries and indexes.  Thank You.